% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quant_Poolingfuns.R
\name{convol.dens}
\alias{convol.dens}
\title{Convolution of two empirical distributions}
\usage{
convol.dens(ecdf1, ecdf2, N1, N2, cutoff)
}
\arguments{
\item{ecdf1}{An ecdf that is the matrix output object from the function \code{convol.dens()} itself or from the function
\code{ecdf_pool()}. The ecdf can be truncated up by \code{cutoff}.}

\item{ecdf2}{An ecdf that is the matrix output object from the function \code{convol.dens()} itself or from the function
\code{ecdf_pool()}. The ecdf can be truncated up by \code{cutoff}.}

\item{N1}{The size of the empirical support of \code{ecdf1}.}

\item{N2}{The size of the empirical support of \code{ecdf2}.}

\item{cutoff}{Cutoff of the support of the resulting convolution distribution}
}
\value{
The function returns a matrix of three columns: The support of convolution distribution
up to \code{cutoff}, empirical PMF and empirical CDF.
}
\description{
The function calculates the convolution of two empirical distributions up to the value of
\code{cutoff} (on the sum of two random variables).
}
\examples{
experiment_1 <- ecdf_pool(c(0,1,2,4,4,4,4,5,0,9,87,12), cutoff = 8)
example_ecdf_pool <- ecdf_pool(round(runif(100, 0, 20)), cutoff = 18)
convol.dens(example_ecdf_pool$out, example_ecdf_pool$out, cutoff = 18)
convol.dens(experiment_1$out, experiment_1$out, cutoff = 8)

}
\references{
Liu T, Hogan JW, Daniels, MJ, Coetzer M, Xu Y, Bove G, et al. Improved HIV-1 Viral Load
Monitoring Capacity Using Pooled Testing with Marker-Assisted Deconvolution. Journal of
AIDS. 2017;75(5): 580-587.
}
\keyword{Pooled}
\keyword{Quantitative}
\keyword{Testing,}
\keyword{convolution,}
\keyword{distribution}
\keyword{empirical}
