% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quant_Poolingfuns.R
\name{ecdf_pool}
\alias{ecdf_pool}
\title{Empirical CDF Calculation for Quantitative Pooled Testing}
\usage{
ecdf_pool(v, N = length(v), cutoff = max(v), ...)
}
\arguments{
\item{v}{Numeric vector of the observations for calculating ecdf. For quantitative pooling
strategies, only CDF on the support of v < cutoff is needed; so \code{v} can be the fraction
of observations less than the cutoff and use \code{N} to pass the number of all observations.}

\item{N}{The number (vector length) of the observations. By default, \code{N = length(v)}.}

\item{cutoff}{Cutoff value of quantitative assay that defines test positivity. By default,
\code{cutoff = max(v)} which calculates the CDF for the entire support of \code{v}.}

\item{...}{Arguments to be passed to subsequent methods.}
}
\value{
The function returns a matrix of three columns: The support of (\code{v <= cutoff}),
empirical PMF and empirical CDF.
}
\description{
Compute an empirical cumulative distribution function, with
results similar to an “ecdf” object.
}
\examples{

ecdf_pool(round(runif(100, 0, 20)), cutoff = 18)

}
\references{
Liu T, Hogan JW, Daniels, MJ, Coetzer M, Xu Y, Bove G, et al. Improved HIV-1 Viral Load
Monitoring Capacity Using Pooled Testing with Marker-Assisted Deconvolution. Journal of
AIDS. 2017;75(5): 580-587.
}
