% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quant_Poolingfuns.R
\name{foo_atr}
\alias{foo_atr}
\title{Intermediate function needed by MP, MiniPooling with Algorithm (MPA), and mMPA}
\usage{
foo_atr(v, N = length(v), max_K = 5, cutoff = 1000, quietly = TRUE)
}
\arguments{
\item{v}{Numeric vector of the observations for calculating ecdf. For quantitative pooling
strategies, only CDF on the support of v < cutoff is needed; so \code{v} can be the fraction
of observations less than the cutoff and use \code{N} to pass the number of all observations.}

\item{N}{The number (vector length) of the observations. By default, \code{N = length(v)}.}

\item{max_K}{Maximum pool size that is considered.}

\item{cutoff}{Cutoff value of quantitative assay that defines test positivity. By default,
\code{cutoff = max(v)} which calculates the CDF for the entire support of \code{v}.}

\item{quietly}{Logical value; whether print the results to screen or not.}
}
\value{
The function returns a matrix of two columns: The pool size from 1:max_K, the convol_cdf(? question about how to describe this value), and the Average Tests Required by MPA (MPA_ATR) given that pool size
}
\description{
This function calculates the ART needed by MPA for a range of pool sizes
(from 2 to a pool size specified).
}
\examples{

foo_atr(c(200, 1500, 1900, 800, 950), max_K = 9, cutoff = 2000, quietly = TRUE)
foo_atr(c(200, 1500, 1900, 800, 950), max_K = 6, cutoff = 1000, quietly = TRUE)

}
\keyword{(?)}
\keyword{(Average}
\keyword{ATR}
\keyword{CDF,}
\keyword{Convolution}
\keyword{Intermediate}
\keyword{MPA_ATR}
\keyword{Pool}
\keyword{Pooled}
\keyword{Quantitative}
\keyword{Required),}
\keyword{Size,}
\keyword{Testing,}
\keyword{Tests}
\keyword{function,}
