% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantPooling_funs.R
\name{mmpa_atr}
\alias{mmpa_atr}
\title{Average Tests Required (ART) needed by Mark-Assisted MiniPooling with Algorithm (mMPA)}
\usage{
mmpa_atr(v, s, N = length(v), max_K = 5, cutoff = 1000,
  quietly = T)
}
\arguments{
\item{v}{Numeric vector of the observations for calculating ecdf. For quantitative pooling
strategies, only CDF on the support of v < cutoff is needed; so \code{v} can be the fraction
of observations less than the cutoff and use \code{N} to pass the number of all observations.}

\item{s}{Risk score that is used for ranked testing}

\item{N}{The number (vector length) of the observations. By default, \code{N = length(v)}.}

\item{max_K}{Maximum pool size that is considered.}

\item{cutoff}{Cutoff value of quantitative assay that defines test positivity. By default,
\code{cutoff = max(v)} which calculates the CDF for the entire support of \code{v}.}

\item{quietly}{Logical value; whether print the results to screen or not.}
}
\description{
This function calculates the ART needed by MP for a range of pool sizes
(from 2 to a pool size specified).
}
