% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quant_Poolingfuns.R
\name{mp_atr}
\alias{mp_atr}
\title{Average Tests Required (ATR) needed by MiniPooling (MP)}
\usage{
mp_atr(v, N = length(v), max_K = 5, cutoff = 1000, quietly = TRUE)
}
\arguments{
\item{v}{Numeric vector of the observations for calculating ecdf. For quantitative pooling
strategies, only CDF on the support of v < cutoff is needed; so \code{v} can be the fraction
of observations less than the cutoff and use \code{N} to pass the number of all observations.}

\item{N}{The number (vector length) of the observations. By default, \code{N = length(v)}.}

\item{max_K}{Maximum pool size that is considered.}

\item{cutoff}{Cutoff value of quantitative assay that defines test positivity. By default,
\code{cutoff = max(v)} which calculates the CDF for the entire support of \code{v}.}

\item{quietly}{Logical value; whether print the results to screen or not.}
}
\value{
The function returns a matrix of two columns: The pool size from 2:max_K and the Average Tests Required by MP given that pool size (MP_ATR)
}
\description{
This function calculates the ART needed by MP for a range of pool sizes
(from 2 to a pool size specified).
}
\examples{

mp_atr(c(200, 1500, 1900, 800, 950), max_K = 9, cutoff = 2000, quietly = TRUE)
mp_atr(c(200, 1500, 1900, 800, 950), max_K = 6, cutoff = 1000, quietly = TRUE)

}
\references{
Liu T, Hogan JW, Daniels, MJ, Coetzer M, Xu Y, Bove G, et al. Improved HIV-1 Viral Load
Monitoring Capacity Using Pooled Testing with Marker-Assisted Deconvolution. Journal of
AIDS. 2017;75(5): 580-587.
}
