% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantPooledTesting-package.R
\docType{package}
\name{QuantPooledTesting-package}
\alias{QuantPooledTesting}
\alias{QuantPooledTesting-package}
\title{QuantPooledTesting: Pooled Testing Strategies with Quantitative Assays}
\description{
The purpose of this package is to calculate the estimated number of quantitative assays that are needed by "mini-pooling", "mini-pooling with algorithm" (May et al. 2010) <doi:10.1097/QAI.0b013e3181ba37a7>, and marker-assisted mini-pooling with algorithm (Liu et al. 2017).Please refer to Liu et al. (2017) <doi:10.1097/QAI.0000000000001424> for more details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/whitneysu/QuantitativePooledTesting}
  \item Report bugs at \url{https://github.com/whitneysu/QuantitativePooledTesting/issues}
}

}
\author{
\strong{Maintainer}: Wanning Su \email{wanning_su@brown.edu}

Authors:
\itemize{
  \item Tao Liu \email{tliu@stat.brown.edu}
}

}
\keyword{internal}
