% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantPooling_funs.R
\name{to_ecdf_class}
\alias{to_ecdf_class}
\title{Convert ecdf_pool output to an \code{ecdf} class}
\usage{
to_ecdf_class(x)
}
\arguments{
\item{x}{Output from the \code{ecdf_pool()} function}
}
\value{
Output in the \code{ecdf} class
}
\description{
Convert ecdf_pool output to an \code{ecdf} class
}
