% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantPooling_funs.R
\name{foo_atr}
\alias{foo_atr}
\title{Intermediat func needed by MiniPooling with Algorithm (MPA)}
\usage{
foo_atr(v, N = length(v), max_K = 5, cutoff = 1000, quietly = F)
}
\arguments{
\item{v}{Numeric vector of the observations for calculating ecdf. For quantitative pooling
strategies, only CDF on the support of v < cutoff is needed; so \code{v} can be the fraction
of observations less than the cutoff and use \code{N} to pass the number of all observations.}

\item{N}{The number (vector length) of the observations. By default, \code{N = length(v)}.}

\item{max_K}{Maximum pool size that is considered.}

\item{cutoff}{Cutoff value of quantitative assay that defines test positivity. By default,
\code{cutoff = max(v)} which calculates the CDF for the entire support of \code{v}.}

\item{quietly}{Logical value; whether print the results to screen or not.}
}
\description{
This function calculates the ART needed by MPA for a range of pool sizes
(from 2 to a pool size specified).
}
