% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quant_Poolingfuns.R
\name{mmpa_atr}
\alias{mmpa_atr}
\title{Average Tests Required (ATR) needed by Marker-Assisted Mini-Pooling with Algorithm (mMPA)}
\usage{
mmpa_atr(v, s, N = length(v), max_K = 5, cutoff = 1000, quietly = TRUE)
}
\arguments{
\item{v}{Numeric vector of the observations for calculating ecdf. For quantitative pooling
strategies, only CDF on the support of v < cutoff is needed; so \code{v} can be the fraction
of observations less than the cutoff and use \code{N} to pass the number of all observations.}

\item{s}{Risk score that is used for ranked testing}

\item{N}{The number (vector length) of the observations. By default, \code{N = length(v)}.}

\item{max_K}{Maximum pool size that is considered.}

\item{cutoff}{Cutoff value of quantitative assay that defines test positivity. By default,
\code{cutoff = max(v)} which calculates the CDF for the entire support of \code{v}.}

\item{quietly}{Logical value; whether print the results to screen or not.}
}
\value{
The function returns a matrix of two columns: The pool size from 2:max_K and the Average Tests Required by mMPA given that pool size (mMPA_ATR)
}
\description{
This function calculates the ATR needed by mMPA for a range of pool sizes
(from 2 to a pool size specified).
}
\examples{

mmpa_atr(c(200, 1500, 1900, 800, 950), s = c(5, 3, 1, 2, 4))

}
\references{
Liu T, Hogan JW, Daniels, MJ, Coetzer M, Xu Y, Bove G, et al. Improved HIV-1 Viral Load
Monitoring Capacity Using Pooled Testing with Marker-Assisted Deconvolution. Journal of
AIDS. 2017;75(5): 580-587. (?)
}
\keyword{(2:K),}
\keyword{(?)}
\keyword{(Average}
\keyword{ATR}
\keyword{Pool}
\keyword{Pooled}
\keyword{Quantitative}
\keyword{Required),}
\keyword{Size}
\keyword{Testing,}
\keyword{Tests}
\keyword{mMPA_ATR}
