% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quant_Poolingfuns.R
\name{to_ecdf_class}
\alias{to_ecdf_class}
\title{Convert ecdf_pool output to an \code{ecdf} class}
\usage{
to_ecdf_class(x)
}
\arguments{
\item{x}{Output object from the \code{ecdf_pool()} function}
}
\value{
Output in the \code{ecdf} class
}
\description{
Convert ecdf_pool output to an \code{ecdf} class
}
\examples{

to_ecdf_class(ecdf_pool(c(0,1,2,4,4,4,4,5,0,9,87,12), cutoff = 8))
to_ecdf_class(ecdf_pool(round(runif(100, 0, 20)), cutoff = 18))

}
\references{
Liu T, Hogan JW, Daniels, MJ, Coetzer M, Xu Y, Bove G, et al. Improved HIV-1 Viral Load
Monitoring Capacity Using Pooled Testing with Marker-Assisted Deconvolution. Journal of
AIDS. 2017;75(5): 580-587.
}
\keyword{Pooled}
\keyword{Quantitative}
\keyword{Testing}
